+++ docs/en/changes/changes.md
@@ -61,6 +61,7 @@
 * Use `groupBy` and `in` operators to optimize topology query for BanyanDB storage plugin.
 * Support server status watcher for `MetricsPersistentWorker` to check the metrics whether required initialization.
 * Fix the meter value are not correct when using `sumPerMinLabeld` or `sumHistogramPercentile` MAL function.
+* Fix cannot display attached events when using Zipkin Lens UI query traces.
 
 #### UI